+++ vehicle/OVMS.V3/components/vehicle_chevrolet_c6_corvette/src/vehicle_chevrolet_c6_corvette.h
@@ -43,7 +43,7 @@ class OvmsVehicleChevroletC6Corvette : public OvmsVehicle
     ~OvmsVehicleChevroletC6Corvette();
 
   protected:
-    void IncomingPollReply(canbus* bus, uint16_t type, uint32_t pid, uint8_t* data, uint8_t length, uint16_t mlremain);
+    void IncomingPollReply(canbus* bus, uint16_t type, uint16_t pid, uint8_t* data, uint8_t length, uint16_t mlremain);
 
   protected:
     char m_vin[18];
