+++ vehicle/OVMS.V3/components/vehicle_mgev/src/mg_can_handler.cpp
@@ -186,7 +186,7 @@ void OvmsVehicleMgEv::IncomingPollFrame(CAN_frame_t* frame)
 }
 
 void OvmsVehicleMgEv::IncomingPollReply(
-        canbus* bus, uint16_t type, uint32_t pid, uint8_t* data, uint8_t length,
+        canbus* bus, uint16_t type, uint16_t pid, uint8_t* data, uint8_t length,
         uint16_t remain)
 {
     ESP_LOGV(
