+++ vehicle/OVMS.V3/components/vehicle_mgev/src/mg_poll_bms.cpp
@@ -101,7 +101,7 @@ void OvmsVehicleMgEv::ProcessBatteryStats(int index, uint8_t* data, uint16_t rem
 }
 
 void OvmsVehicleMgEv::IncomingBmsPoll(
-        uint32_t pid, uint8_t* data, uint8_t length, uint16_t remain)
+        uint16_t pid, uint8_t* data, uint8_t length, uint16_t remain)
 {
     uint16_t value = (data[0] << 8 | data[1]);
