+++ vehicle/OVMS.V3/components/vehicle_voltampera/src/vehicle_voltampera.h
@@ -87,7 +87,7 @@ class OvmsVehicleVoltAmpera : public OvmsVehicle
     void IncomingFrameCan2(CAN_frame_t* p_frame);
     void IncomingFrameCan3(CAN_frame_t* p_frame);
     void IncomingFrameCan4(CAN_frame_t* p_frame);
-    void IncomingPollReply(canbus* bus, uint16_t type, uint32_t pid, uint8_t* data, uint8_t length, uint16_t mlremain);
+    void IncomingPollReply(canbus* bus, uint16_t type, uint16_t pid, uint8_t* data, uint8_t length, uint16_t mlremain);
     void TxCallback(const CAN_frame_t* p_frame, bool success);
     void CommandWakeupComplete(const CAN_frame_t* p_frame, bool success);
     void SendTesterPresentMessage( uint32_t id );
